copy over token separation logic from older PR
